requirements.txt
Flask
Flask-SocketIO
PyYAML

